add helper class
