+++ core/src/main/java/org/apache/gravitino/Entity.java
@@ -117,29 +117,6 @@ public interface Entity extends Serializable {
   interface RelationalEntity<E extends Entity & HasIdentifier> {
 
     /**
-     * Creates a new instance of {@link RelationalEntity}.
-     *
-     * @param entity the entity to be wrapped
-     * @param vertexType the type of the vertex in the relation
-     * @param relatedIdents the list of related identifiers
-     * @param relatedEntityType the type of the related entity
-     * @return a new instance of {@link RelationalEntity}
-     * @param <E> the type of the entity, which must extend {@link Entity} and implement {@link
-     *     HasIdentifier}
-     */
-    static <E extends Entity & HasIdentifier> RelationalEntity<E> of(
-        E entity,
-        Relation.VertexType vertexType,
-        List<NameIdentifier> relatedIdents,
-        EntityType relatedEntityType) {
-      if (vertexType == Relation.VertexType.SOURCE) {
-        return new SourceRelationEntity<>(entity, relatedIdents, relatedEntityType);
-      } else {
-        return new DestRelationEntity<>(entity, relatedIdents, relatedEntityType);
-      }
-    }
-
-    /**
      * Returns the entity wrapped by this relational entity.
      *
      * @return the entity
@@ -166,87 +143,5 @@ public interface Entity extends Serializable {
      * @return the type of the related entity
      */
     EntityType relatedEntityType();
-  }
-
-  /**
-   * Represents a relational entity that contains a source vertex entity, its vertex type, related
-   * destination vertex identifiers, and the type of the destination entity.
-   *
-   * @param <E> the type of the entity, which must extend {@link Entity} and implement {@link
-   *     HasIdentifier}
-   */
-  class SourceRelationEntity<E extends Entity & HasIdentifier> implements RelationalEntity<E> {
-    final E sourceVertexEntity;
-    final List<NameIdentifier> destVertexIdents;
-    final EntityType destVertexEntityType;
-
-    SourceRelationEntity(
-        E sourceVertexEntity,
-        List<NameIdentifier> destVertexIdents,
-        EntityType destVertexEntityType) {
-      this.sourceVertexEntity = sourceVertexEntity;
-      this.destVertexIdents = destVertexIdents;
-      this.destVertexEntityType = destVertexEntityType;
-    }
-
-    @Override
-    public E entity() {
-      return sourceVertexEntity;
-    }
-
-    @Override
-    public Relation.VertexType vertexType() {
-      return Relation.VertexType.SOURCE;
-    }
-
-    @Override
-    public List<NameIdentifier> relatedNameIdentifiers() {
-      return destVertexIdents;
-    }
-
-    @Override
-    public EntityType relatedEntityType() {
-      return destVertexEntityType;
-    }
-  }
-
-  /**
-   * Represents a relational entity that contains a destination vertex entity, its vertex type,
-   * related source vertex identifiers, and the type of the source entity.
-   *
-   * @param <E> the type of the entity, which must extend {@link Entity} and implement {@link
-   *     HasIdentifier}
-   */
-  class DestRelationEntity<E extends Entity & HasIdentifier> implements RelationalEntity<E> {
-    final E destVertexEntity;
-    final List<NameIdentifier> sourceVertexIdents;
-    final EntityType sourceVertexEntityType;
-
-    DestRelationEntity(
-        E destVertexEntity, List<NameIdentifier> sourceVertexIdents, EntityType sourceEntityType) {
-      this.destVertexEntity = destVertexEntity;
-      this.sourceVertexIdents = sourceVertexIdents;
-      this.sourceVertexEntityType = sourceEntityType;
-    }
-
-    @Override
-    public E entity() {
-      return destVertexEntity;
-    }
-
-    @Override
-    public Relation.VertexType vertexType() {
-      return Relation.VertexType.DESTINATION;
-    }
-
-    @Override
-    public List<NameIdentifier> relatedNameIdentifiers() {
-      return sourceVertexIdents;
-    }
-
-    @Override
-    public EntityType relatedEntityType() {
-      return sourceVertexEntityType;
-    }
   }
 }
